feat: Add Vorze A10 Cyclone Support
Add support for device finding and sending proprietary commands via
Bluetooth LE. Still need to implement USB dongle support.

Fixes #70

+++ Buttplug/Bluetooth/Devices/FleshlightLaunch.cs
@@ -40,10 +40,10 @@ namespace Buttplug.Bluetooth.Devices
     internal class FleshlightLaunch : ButtplugBluetoothDevice
     {
         private readonly Stopwatch _stopwatch;
-        private readonly ushort _previousSpeed;
-        private readonly ushort _previousPosition;
-        private ushort _previousKiirooPosition;
-        private ushort _limitedSpeed;
+        private readonly uint _previousSpeed;
+        private readonly uint _previousPosition;
+        private uint _previousKiirooPosition;
+        private uint _limitedSpeed;
 
         public FleshlightLaunch(IButtplugLogManager aLogManager,
                                 IBluetoothDeviceInterface aInterface) :
